2dcontext/scroll/scrollPathIntoView.html

<!DOCTYPE html>
<html>
<head>
<meta charset="UTF-8">
<title>scrollPathIntoView</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
</head>
<body>
<div id="log"></div>
<canvas id="c" width="300" height="150" style="position: absolute; top: 100px; left: 200px;"></canvas>
<div style="width: 200vw; height: 200vh;"></div>
<script>
var t = async_test('');
_addTest(function (canvas, ctx) {
    window.scrollTo(0, 0);
    ctx.beginPath();
    ctx.rect(10, 20, 40, 80);
    ctx.scrollPathIntoView();
    var rect = c.getBoundingClientRect();
    assert_approx_equals(rect.top, -20, 0.5, 'The canvas top edge is 20px above the viewport top edge');
    assert_approx_equals(rect.left, 200, 0.5, 'The canvas left edge is on 200px right of the viewport left edge');
});
</script>
</body>
</html>
